Fix accidental data accumulation (#470)
Previously, the `upsertResources` function could repeatedly decode items
into the same resource, and then insert that updated resource into the
database. Fields from the new item would overwrite existing fields in
the resource, but it would not remove the fields that were not set in
the new item, causing incorrect data to accumulate in some resources.

Now the resource is reset before decoding.

Refs:
 - https://issues.redhat.com/browse/ACM-21047

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>